refactor: move ProjectDenied to organisms
