+++ web-registry/src/components/organisms/ProjectDenied/ProjectDenied.tsx
